refactored part of the command callbacks into their own classes MinifiableData fixes and improvements refactored "commands" to "help"
