+++ grocy_telegram_bot/commands/stats.py
@@ -1,0 +1,38 @@
+from telegram import Update
+from telegram.ext import Filters, CommandHandler, CallbackContext
+from telegram_click.decorator import command
+
+from grocy_telegram_bot.commands import GrocyCommandHandler
+from grocy_telegram_bot.const import COMMAND_STATS
+from grocy_telegram_bot.permissions import CONFIG_ADMINS
+from grocy_telegram_bot.stats import format_metrics
+from grocy_telegram_bot.util import send_message
+
+
+class StatsCommandHandler(GrocyCommandHandler):
+
+    def command_handlers(self):
+        return [
+            CommandHandler(COMMAND_STATS,
+                           filters=(~ Filters.reply) & (~ Filters.forwarded),
+                           callback=self._stats_callback),
+        ]
+
+    @command(
+        name=COMMAND_STATS,
+        description="List statistics of this bot.",
+        permissions=CONFIG_ADMINS
+    )
+    def _stats_callback(self, update: Update, context: CallbackContext) -> None:
+        """
+        /stats command handler
+        :param update: the chat update object
+        :param context: telegram context
+        """
+        bot = context.bot
+        message = update.effective_message
+        chat_id = update.effective_chat.id
+
+        text = format_metrics()
+
+        send_message(bot, chat_id, text, reply_to=message.message_id)
